streaming/producers/requirements.txt
kafka-python==2.0.2
faker==19.6.2
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0